backend/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
celery[redis]
requests
python-multipart
numpy
transformers>=4.28.0
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
pinecone-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-celery
prometheus-client
pytest
docker
moviepy
gTTS
python-pptx
beautifulsoup4
aioredis
sqlmodel
alembic
Pillow
asyncpg
sqlalchemy[asyncio]
httpx[http2]
slowapi
starlette-helmet
pytest-asyncio
sentry-sdk
python-json-logger
kubernetes
pytest-xdist
